requirements.txt
minigrid==2.3.1
anthropic==0.34.2
openai==1.97.1
scale-llm-engine=0.0.0beta38
yaml
nltk
